+++ Utilities/AnalysisUtilities.h
@@ -4,15 +4,15 @@
 // Analysis Utilities
 void NormaliseYieldToNEntries(TH1D* histogram);
 void NormaliseRawHistToNEntries(TH1D* histogram);
-void NormaliseYieldToNEvents(TH1D* histogram, int nEvents);
-void NormaliseRawHistToNEvents(TH1D* histogram, int nEvents);
+void NormaliseYieldToNEvents(TH1D* histogram, double nEvents);
+void NormaliseRawHistToNEvents(TH1D* histogram, double nEvents);
 void NormaliseYieldToIntegral(TH1D* histogram);
 void NormaliseRawHistToIntegral(TH1D* histogram);
 int GetNEvents(TFile* file_O2Analysis, TString analysisWorkflow);
 int GetNEventsGen(TFile* file_O2Analysis);
 int GetNEventsSelected_JetFramework(TFile* file_O2Analysis);
 int GetNEventsSelected_TrackEffWorkflow(TFile* file_O2Analysis);
-int GetNEventsSelected_JetFramework_weighted(TFile* file_O2Analysis);
+double GetNEventsSelected_JetFramework_weighted(TFile* file_O2Analysis);
 int GetNEventsSel8Centrality(TFile* file_O2Analysis, float centralityLow, float centralityHigh);
 int GetNEventsSelectedCentrality_JetFramework(TFile* file_O2Analysis, float centralityLow, float centralityHigh, const char trainId[]);
